# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance and productivity API",
        "title": "Kaizen API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@kaizen.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/journals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all journal entries with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "List journal entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type (income or expense)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.JournalListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new income or expense entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "Create a journal entry",
                "parameters": [
                    {
                        "description": "Journal entry details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateJournalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceJournal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/journals/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get income, expense, and balance summary for a date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "Get financial summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD, default: first day of current month)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD, default: today)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.JournalSummary"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/journals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single journal entry by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "Get a journal entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceJournal"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing journal entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "Update a journal entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Journal data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateJournalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceJournal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a journal entry (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Journals"
                ],
                "summary": "Delete a journal entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Journal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jedi116_kaizen-api_internal_models.FinanceJournal": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Transaction amount (always positive)",
                    "type": "number"
                },
                "category": {
                    "description": "Belongs to a category",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                        }
                    ]
                },
                "category_id": {
                    "description": "Which category (Food, Rent, etc.)",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "description": "Transaction date (for daily tracking)",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "Optional detailed notes",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_recurring": {
                    "description": "Is this a recurring transaction?",
                    "type": "boolean"
                },
                "location": {
                    "description": "Where transaction occurred (optional)",
                    "type": "string"
                },
                "payment_method": {
                    "description": "\"cash\", \"credit_card\", \"bank_transfer\", etc.",
                    "type": "string"
                },
                "receipt_url": {
                    "description": "Attachments (optional - for receipts)",
                    "type": "string"
                },
                "title": {
                    "description": "e.g., \"Grocery shopping at Walmart\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"income\" or \"expense\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Which user made this transaction",
                    "type": "integer"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers.CreateJournalRequest": {
            "type": "object",
            "required": [
                "amount",
                "category_id",
                "title"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 25.5
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "date": {
                    "type": "string",
                    "example": "2025-01-15"
                },
                "description": {
                    "type": "string",
                    "example": "Weekly groceries at Walmart"
                },
                "is_recurring": {
                    "type": "boolean",
                    "example": false
                },
                "location": {
                    "type": "string",
                    "example": "Walmart"
                },
                "payment_method": {
                    "type": "string",
                    "example": "credit_card"
                },
                "receipt_url": {
                    "type": "string",
                    "example": "https://example.com/receipt.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Grocery shopping"
                }
            }
        },
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "internal_handlers.JournalListResponse": {
            "type": "object",
            "properties": {
                "journals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceJournal"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "internal_handlers.JournalSummary": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2025-01-31"
                },
                "entry_count": {
                    "type": "integer",
                    "example": 42
                },
                "net_balance": {
                    "type": "number",
                    "example": 1500
                },
                "start_date": {
                    "type": "string",
                    "example": "2025-01-01"
                },
                "total_expense": {
                    "type": "number",
                    "example": 3500
                },
                "total_income": {
                    "type": "number",
                    "example": 5000
                }
            }
        },
...
        "internal_handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "internal_handlers.UpdateJournalRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 25.5
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "date": {
                    "type": "string",
                    "example": "2025-01-15"
                },
                "description": {
                    "type": "string",
                    "example": "Weekly groceries at Walmart"
                },
                "is_recurring": {
                    "type": "boolean",
                    "example": false
                },
                "location": {
                    "type": "string",
                    "example": "Walmart"
                },
                "payment_method": {
                    "type": "string",
                    "example": "credit_card"
                },
                "receipt_url": {
                    "type": "string",
                    "example": "https://example.com/receipt.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Grocery shopping"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}